Preserve search parameters when going to the details page
When clicking on a link to a details page, we preserve the URL parameters in order. This allows the user to refresh page and still be able to see the pagination on details pages.

It works by saving the current URL parameters, along with the route path and the API.

This fixes #45